Change entry point to main.rs, and cli to cli.rs
